# Backend/src/api/openapi.yaml
...
openapi: 3.0.2
info:
  title: Kochbuch API
  version: 1.0.0
  description: |-
    Die Kochbuch API ist eine RESTful API, die zwei Microservices bereitstellt, um eine 2-Page-App für die Verwaltung von Kochrezepten zu unterstützen. Der erste Microservice bietet Funktionen zur Verwaltung von Benutzerkonten, einschließlich der Registrierung, Authentifizierung und Autorisierung von Benutzern. Der zweite Microservice verwaltet Kochrezepte und bietet Funktionen zum Erstellen, Bearbeiten, Löschen und Anzeigen von Rezepten.

    Die API bietet eine Reihe von Endpunkten, die es der 2-Page-App ermöglichen, auf die Funktionen der beiden Microservices zuzugreifen. Die Endpunkte für den Benutzerservice umfassen die Registrierung eines neuen Benutzers, die Anmeldung eines bestehenden Benutzers, die Rückgabe von Benutzerinformationen und die Verwaltung von Berechtigungen.
  contact:
    name: "Moritz Rau, Jan-Niklas Erdmann"
    email: moritz_rau@web.de
  license:
    name: The Unlicense
    url: http://unlicense.org
paths:
...
  /login/{loginId}:
    summary: Path used to manage a single login.
    description: "The REST endpoint/path used to get, update, and delete single instances\
      \ of an `login`.  This path contains `GET`, `PUT`, and `DELETE` operations used\
      \ to perform the get, update, and delete tasks, respectively."
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/login'
          description: Successful response - returns a single `login`.
      operationId: getlogin
      summary: Get a login
      description: Gets the details of a single instance of a `login`.
    put:
      requestBody:
        description: Updated `login` information.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
        required: true
      responses:
        "202":
          description: Successful response.
      operationId: updatelogin
      summary: Update a login
      description: Updates an existing `login`.
    delete:
      responses:
        "204":
          description: Successful response.
      operationId: deletelogin
      summary: Delete a login
      description: Deletes an existing `login`.
    parameters:
    - name: loginId
      description: A unique identifier for a `login`.
      schema:
        type: string
      in: path
      required: true
components:
  schemas:
    login:
      title: Root Type for login
      description: Login-Vorgang für Verwender des Kochbuchs
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        _links:
          $ref: '#/components/schemas/HATEOAS_Resource'
          description: ""
      example:
        username: frederike
        email: mulder@xfiles.com
        password: Bratwurst123456789$
...
    HATEOS_Link:
      title: Root Type for HATEOS_Link
      description: HATEOS-Link
      type: object
      properties:
        url:
          type: string
        method:
          type: string
        query_params:
          type: array
          items:
            type: string
      example:
        url: /Recipe
        method: GET
        query_params:
        - name
        - difficulty
        - time
        - ingredients
        - description
    HATEOAS_Resource:
      title: Root Type for HATEOAS_Resource
      description: HATEOAS-Links einer Resource
      type: object
      properties:
        read:
          $ref: '#/components/schemas/HATEOS_Link'
        update:
          $ref: '#/components/schemas/HATEOS_Link'
        patch:
          $ref: '#/components/schemas/HATEOS_Link'
        delete:
          $ref: '#/components/schemas/HATEOS_Link'
      example:
        read: ""
        update: ""
        patch: ""
        delete: ""
...